+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+- Added `BinaryTypes.declare()`
+- Added `BinaryTypes.typeDef()`
+- Removed `BinaryTypeHelper.typeDef()`
+
 ## 0.0.3
 
 - Changes in the `PointerType.name`. Array types quoted for the better readability. Eg. `(char[10])*` means a pointer to `char[10]`.